Update dependencies of app.js move
