+++ src/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./styles/index.css";
-import App from "./components/App.js";
+import App from "./App.js";
 
 ReactDOM.render(<App />, document.getElementById("root"));
